pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fsc-assistant"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Sunny Liu", email = "sunnyliu2@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.3.0",
    "docling>=2.58.0",
    "openai>=2.6.1",
    "pillow>=10.4.0",
    "prompt-toolkit>=3.0.52",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "tomlkit>=0.13.3",
    "certifi>=2025.10.5",
    "pygithub>=2.8.1",
    "jira>=3.10.5",
    "fastmcp>=2.13.0.2",
]


# [tool.uv.sources]
# torch = { index = "pytorch" }
# torchvision = { index = "pytorch" }
# torchaudio = { index = "pytorch" }


# [[tool.uv.index]]
# name = "default"
# url = "https://pypi.org/simple"
# default = true

# [[tool.uv.index]]
# name = "pytorch"
# url = "https://download.pytorch.org/whl/cu118"


[tool.setuptools.packages.find]
where = ["./src"]

[tool.setuptools.package-data]
"fsc" = ["**/*"]

[project.scripts]
fsc = "fsc.cli:cli"

[project.urls]
Homepage = "https://github.com/devsunny/fsc-assistant"
Documentation = "https://github.com/devsunny/fsc-assistant/README.md"
Repository = "https://github.com/devsunny/fsc-assistant"
